Adição de listagem de resgates
